Automatically determine transaction concurrency
